Move more functionality to UpdatingElement
* Support for renderRoot and static styles moved from LitElement to UpdatingElement
* All decorators moved to updating-element package. Note, they are re-exported in lit-element
* LitElement 'lib' folder removed. Decorators are now exported at `lit-element/decorators/*`.

(WIP) platform-support is still entirely handled in LitElement. The non-lit portion should move to UpdatingElement.
