+++ packages/lit-element/rollup.config.js
@@ -19,18 +19,16 @@ import {
 export default litRollupConfig({
   entryPoints: [
     'lit-element',
-    'lib/platform-support',
-    'lib/css-tag',
-    'lib/decorators',
-    'lib/decorators/base',
-    'lib/decorators/customElement',
-    'lib/decorators/eventOptions',
-    'lib/decorators/internalProperty',
-    'lib/decorators/property',
-    'lib/decorators/query',
-    'lib/decorators/queryAll',
-    'lib/decorators/queryAssignedNodes',
-    'lib/decorators/queryAsync',
+    'platform-support',
+    'decorators',
+    'decorators/customElement',
+    'decorators/eventOptions',
+    'decorators/internalProperty',
+    'decorators/property',
+    'decorators/query',
+    'decorators/queryAll',
+    'decorators/queryAssignedNodes',
+    'decorators/queryAsync',
   ],
   external: ['lit-html', 'updating-element']
 });
