+++ packages/lit-element/src/decorators.ts
