+++ packages/lit-element/src/decorators/customElement.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright (c) 2020 The Polymer Project Authors. All rights reserved.
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
  * This code may only be used under the BSD style license found at
  * http://polymer.github.io/LICENSE.txt
  * The complete set of authors may be found at
@@ -11,7 +11,5 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import '../../platform-support.js';
-import '../lit-element_styling_test.js';
-import '../lit-element_styling_apply_test.js';
-import '../lit-element_test.js';
+
+export * from 'updating-element/decorators/customElement.js';
