+++ packages/lit-element/src/decorators/internalProperty.ts
