+++ packages/lit-element/src/decorators/property.ts
