+++ packages/lit-element/src/test/platform-support/lit-html-scoping-shim_html-test.ts
@@ -11,8 +11,8 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import '../../lib/platform-support.js';
-import {renderShadowRoot} from '../test-utils/shadow-root';
+import '../../platform-support.js';
+import {renderShadowRoot} from '../test-helpers.js';
 import {html} from 'lit-html';
 import {assert} from '@esm-bundle/chai';
