+++ packages/lit-element/src/test/platform-support/lit-html_html-test.ts
@@ -11,10 +11,10 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import '../../lib/platform-support.js';
+import '../../platform-support.js';
 import {html} from 'lit-html';
 import {assert} from '@esm-bundle/chai';
-import {renderShadowRoot} from '../test-utils/shadow-root';
+import {renderShadowRoot} from '../test-helpers.js';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
