+++ packages/lit-element/tsconfig.json
