+++ packages/tests/web-test-runner.config.js
@@ -160,6 +160,7 @@ export default {
   files: [
     '../lit-html/development/**/*_test.js',
     '../lit-element/development/**/*_test.(js|html)',
+    '../updating-element/development/**/*_test.(js|html)',
   ],
   nodeResolve: true,
   concurrency: 6, // default cores / 2
