+++ packages/updating-element/.gitignore
@@ -1,3 +1,5 @@
 /development/
 /decorators/
 /updating-element.*
+/css-tag.*
+/decorators.*
