+++ packages/updating-element/src/decorators.ts
@@ -22,3 +22,8 @@
 export * from './decorators/customElement.js';
 export * from './decorators/property.js';
 export * from './decorators/internalProperty.js';
+export * from './decorators/eventOptions.js';
+export * from './decorators/query.js';
+export * from './decorators/queryAll.js';
+export * from './decorators/queryAsync.js';
+export * from './decorators/queryAssignedNodes.js';
