+++ packages/updating-element/src/decorators/queryAll.ts
@@ -19,35 +19,27 @@
  * not an arrow function.
  */
 
-import {LitElement} from '../../lit-element.js';
+import {UpdatingElement} from '../updating-element.js';
 import {
   ClassElement,
   legacyPrototypeMethod,
   standardPrototypeMethod,
 } from './base.js';
 
-// Note, in the future, we may extend this decorator to support the use case
-// where the queried element may need to do work to become ready to interact
-// with (e.g. load some implementation code). If so, we might elect to
-// add a second argument defining a function that can be run to make the
-// queried element loaded/updated/ready.
 /**
- * A property decorator that converts a class property into a getter that
- * returns a promise that resolves to the result of a querySelector on the
- * element's renderRoot done after the element's `updateComplete` promise
- * resolves. When the queried property may change with element state, this
- * decorator can be used instead of requiring users to await the
- * `updateComplete` before accessing the property.
+ * A property decorator that converts a class property into a getter
+ * that executes a querySelectorAll on the element's renderRoot.
  *
  * @param selector A DOMString containing one or more selectors to match.
  *
- * See: https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector
+ * See:
+ * https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll
  *
  * @example
  * ```ts
  * class MyElement {
- *   @queryAsync('#first')
- *   first;
+ *   @queryAll('div')
+ *   divs;
  *
  *   render() {
  *     return html`
@@ -56,24 +48,18 @@ import {
  *     `;
  *   }
  * }
- *
- * // external usage
- * async doSomethingWithFirst() {
- *  (await aMyElement.first).doSomething();
- * }
  * ```
  * @category Decorator
  */
-export function queryAsync(selector: string) {
+export function queryAll(selector: string) {
   return (
     protoOrDescriptor: Object | ClassElement,
     name?: PropertyKey
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ): any => {
     const descriptor = {
-      async get(this: LitElement) {
-        await this.updateComplete;
-        return this.renderRoot.querySelector(selector);
+      get(this: UpdatingElement) {
+        return this.renderRoot?.querySelectorAll(selector);
       },
       enumerable: true,
       configurable: true,
