+++ packages/updating-element/src/decorators/queryAssignedNodes.ts
@@ -19,70 +19,75 @@
  * not an arrow function.
  */
 
-import {LitElement} from '../../lit-element.js';
+import {UpdatingElement} from '../updating-element.js';
 import {
   ClassElement,
   legacyPrototypeMethod,
   standardPrototypeMethod,
 } from './base.js';
 
+// x-browser support for matches
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const ElementProto = Element.prototype as any;
+const legacyMatches =
+  ElementProto.msMatchesSelector || ElementProto.webkitMatchesSelector;
+
 /**
  * A property decorator that converts a class property into a getter that
- * executes a querySelector on the element's renderRoot.
+ * returns the `assignedNodes` of the given named `slot`. Note, the type of
+ * this property should be annotated as `NodeListOf<HTMLElement>`.
  *
- * @param selector A DOMString containing one or more selectors to match.
- * @param cache An optional boolean which when true performs the DOM query only
- * once and caches the result.
+ * @param slotName A string name of the slot.
+ * @param flatten A boolean which when true flattens the assigned nodes,
+ * meaning any assigned nodes that are slot elements are replaced with their
+ * assigned nodes.
+ * @param selector A string which filters the results to elements that match
+ * the given css selector.
  *
- * See: https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector
- *
- * @example
- *
+ * * @example
  * ```ts
  * class MyElement {
- *   @query('#first')
- *   first;
+ *   @queryAssignedNodes('list', true, '.item')
+ *   listItems;
  *
  *   render() {
  *     return html`
- *       <div id="first"></div>
- *       <div id="second"></div>
+ *       <slot name="list"></slot>
  *     `;
  *   }
  * }
  * ```
  * @category Decorator
  */
-export function query(selector: string, cache?: boolean) {
+export function queryAssignedNodes(
+  slotName = '',
+  flatten = false,
+  selector = ''
+) {
   return (
     protoOrDescriptor: Object | ClassElement,
     name?: PropertyKey
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ): any => {
     const descriptor = {
-      get(this: LitElement) {
-        return this.renderRoot.querySelector(selector);
+      get(this: UpdatingElement) {
+        const slotSelector = `slot${
+          slotName ? `[name=${slotName}]` : ':not([name])'
+        }`;
+        const slot = this.renderRoot?.querySelector(slotSelector);
+        let nodes = (slot as HTMLSlotElement)?.assignedNodes({flatten});
+        if (nodes && selector) {
+          nodes = nodes.filter((node) =>
+            node.nodeType === Node.ELEMENT_NODE && (node as Element).matches
+              ? (node as Element).matches(selector)
+              : legacyMatches.call(node as Element, selector)
+          );
+        }
+        return nodes;
       },
       enumerable: true,
       configurable: true,
     };
-    if (cache) {
-      const key = typeof name === 'symbol' ? Symbol() : `__${name}`;
-      descriptor.get = function (this: LitElement) {
-        if (
-          ((this as unknown) as {[key: string]: Element | null})[
-            key as string
-          ] === undefined
-        ) {
-          ((this as unknown) as {[key: string]: Element | null})[
-            key as string
-          ] = this.renderRoot.querySelector(selector);
-        }
-        return ((this as unknown) as {[key: string]: Element | null})[
-          key as string
-        ];
-      };
-    }
     return name !== undefined
       ? legacyPrototypeMethod(descriptor, protoOrDescriptor as Object, name)
       : standardPrototypeMethod(descriptor, protoOrDescriptor as ClassElement);
