+++ packages/updating-element/src/test/decorators/queryAll_test.ts
@@ -12,40 +12,40 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {html, LitElement} from '../../lit-element.js';
-import {queryAsync} from '../../lib/decorators/queryAsync.js';
-import {canTestLitElement, generateElementName} from '../test-helpers.js';
+import {queryAll} from '../../decorators/queryAll.js';
+import {
+  canTestUpdatingElement,
+  generateElementName,
+  RenderingElement,
+  html,
+} from '../test-helpers.js';
 import {assert} from '@esm-bundle/chai';
 
-(canTestLitElement ? suite : suite.skip)('@queryAsync', () => {
+(canTestUpdatingElement ? suite : suite.skip)('@queryAll', () => {
   let container: HTMLElement;
   let el: C;
 
-  class C extends LitElement {
-    @queryAsync('#blah') blah!: Promise<HTMLDivElement>;
-    @queryAsync('span') nope!: Promise<HTMLSpanElement | null>;
+  class C extends RenderingElement {
+    @queryAll('div') divs!: NodeList;
 
-    foo = false;
-
-    static properties = {foo: {}};
+    @queryAll('span') spans!: NodeList;
 
     render() {
       return html`
         <div>Not this one</div>
-        ${this.foo
-          ? html`<div id="blah" foo>This one</div>`
-          : html`<div id="blah">This one</div>`}
+        <div id="blah">This one</div>
       `;
     }
   }
   customElements.define(generateElementName(), C);
 
-  setup(() => {
+  setup(async () => {
     container = document.createElement('div');
     container.id = 'test-container';
     document.body.appendChild(container);
     el = new C();
     container.appendChild(el);
+    await el.updateComplete;
   });
 
   teardown(() => {
@@ -55,18 +55,19 @@ import {assert} from '@esm-bundle/chai';
     }
   });
 
-  test('returns an element when it exists after update', async () => {
-    let div = await el.blah;
-    assert.instanceOf(div, HTMLDivElement);
-    assert.isFalse(div.hasAttribute('foo'));
-    el.foo = true;
-    div = await el.blah;
-    assert.instanceOf(div, HTMLDivElement);
-    assert.isTrue(div.hasAttribute('foo'));
+  test('returns elements when they exists', () => {
+    assert.lengthOf(el.divs, 2);
+    assert.deepEqual(
+      Array.from(el.divs),
+      Array.from(el.renderRoot.querySelectorAll('div'))
+    );
   });
 
-  test('returns null when no match', async () => {
-    const span = await el.nope;
-    assert.isNull(span);
+  test('returns empty NodeList when no match', () => {
+    assert.lengthOf(el.spans, 0);
+    assert.deepEqual(
+      Array.from(el.spans),
+      Array.from(el.renderRoot.querySelectorAll('span'))
+    );
   });
 });
