+++ packages/updating-element/src/test/decorators/query_test.ts
@@ -12,16 +12,20 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {html, LitElement} from '../../lit-element.js';
-import {query} from '../../lib/decorators/query.js';
-import {canTestLitElement, generateElementName} from '../test-helpers.js';
+import {query} from '../../decorators/query.js';
+import {
+  canTestUpdatingElement,
+  generateElementName,
+  RenderingElement,
+  html,
+} from '../test-helpers.js';
 import {assert} from '@esm-bundle/chai';
 
-(canTestLitElement ? suite : suite.skip)('@query', () => {
+(canTestUpdatingElement ? suite : suite.skip)('@query', () => {
   let container: HTMLElement;
   let el: C;
 
-  class C extends LitElement {
+  class C extends RenderingElement {
     @query('#blah') div?: HTMLDivElement;
     @query('span', true) span?: HTMLSpanElement;
